+++ data-raw/to_update/repackage_krapp2021.R
@@ -41,11 +41,22 @@ if (!dir.exists("./repackaged")){
 if (!dir.exists("./temp")){
   dir.create("./temp")
 }
+
+# now process the biomes
+if (!dir.exists("./original_files/vegetation_annual")){
+  dir.create("./original_files/vegetation_annual")
+}
+if (file.exists("./original_files/vegetation/biome4output_800ka.nc")){
+  file.copy("./original_files/vegetation/biome4output_800ka.nc",
+            "./original_files/vegetation_annual/biome4output_800ka.nc")
+  file.remove("./original_files/vegetation/biome4output_800ka.nc")
+}
+
 # create landmask
-cdo("select,name=biome ./original_files/vegetation/biome4output_800ka.nc krapp_biome.nc")
-cdo ("-O -expr,'biome = ((biome < 28)) ? biome : -1' krapp_biome.nc krapp_biome2.nc")
-cdo ("-O -expr,'biome = ((biome >= 0)) ? 1 : 0' krapp_biome2.nc krapp_land_only.nc")
-file.remove("krapp_biome.nc", "krapp_biome2.nc")
+cdo("select,name=biome ./original_files/vegetation_annual/biome4output_800ka.nc ./temp/krapp_biome.nc")
+cdo ("-O -expr,'biome = ((biome < 28)) ? biome : -1' ./temp/krapp_biome.nc ./temp/krapp_biome2.nc")
+cdo ("-O -expr,'biome = ((biome >= 0)) ? 1 : 0' ./temp/krapp_biome2.nc ./temp/krapp_land_only.nc")
+file.remove("./temp/krapp_biome.nc", "./temp/krapp_biome2.nc")
 
 # now process each file
 download_dir <- file.path(wkdir,"original_files/bioclimate")
@@ -64,21 +75,15 @@ for (file_id in list.files(download_dir,full.names = TRUE)){
   new_file_cut <- file.path("./repackaged/cut", paste0("Krapp2021_",var1,"_v",file_version,".nc"))
   new_file_uncut <- file.path("./repackaged/uncut", paste0("Krapp2021_uncut_",var1,"_v",file_version,".nc"))
   
-  cdo(paste0("-z zip_9 div ",file_id," krapp_land_only.nc ",new_file_cut))
-  file.copy(from= file_id, to=new_file_uncut)
+  cdo(paste0("-z zip_9 div ",file_id," ./temp/krapp_land_only.nc ",new_file_cut))
+  
+  ## TODO this needs to be reinstated to create uncut files
+  #file.copy(from= file_id, to=new_file_uncut)
   # NOTE that the above does not recompress the file
   #cdo(paste0("-z zip_9 ",file_id," ",new_file_uncut))
 }
 
-# now process the biomes
-if (!dir.exists("./original_files/vegetation_annual")){
-  dir.create("./original_files/vegetation_annual")
-}
-if (file.exists("./original_files/vegetation/biome4output_800ka.nc")){
-  file.copy("./original_files/vegetation/biome4output_800ka.nc",
-            "./original_files/vegetation_annual/biome4output_800ka.nc")
-  file.remove("./original_files/vegetation/biome4output_800ka.nc")
-}
+
   ##TODO this has not been implemented yet!!!!
   #new_file <- file.path("./repackaged/", paste0(var1,"_biome_v",file_version,".nc"))
   #ncks(paste0("-C -O -x -v pco2,npp,LAI ", file_id, " ", new_file))
@@ -99,52 +104,56 @@ monthly_dirs <- file.path(wkdir,"original_files", monthly_dirs)
 # 
 
 
-##TODO 
-#We need a temp/monthly directory
+# ##TODO 
+# #We need a temp/monthly directory
+# 
+# # use the landmask to cut out the ice
+# month_prefix<-tolower(substr(month.name,1,3))
+# for (this_file in list.files(monthly_dirs, full.names = TRUE)){
+#   i <- basename(this_file)
+#   # get the month
+#   this_month<-unlist(strsplit(unlist(strsplit(i,"_"))[3],".",fixed=T))[1]
+#   this_month_id <- match(this_month,month_prefix)
+#   if (this_month_id<10){
+#     this_month_id <- paste0("0",this_month_id)
+#   }
+#   new_filename<-paste(unlist(strsplit(i,"_"))[1:2],collapse="_")
+#   new_filename<-paste0(new_filename,"_",this_month_id,".nc")
+#   new_filename<-file.path("./temp/monthly",new_filename)
+#   # mask the file
+#   if (unlist(strsplit(i,"_"))[1]=="biome4output"){
+#     cdo ("-z zip_9 div -selname,mo_npp",this_file,
+#          " krapp_land_only.nc ", new_filename)} else
+#          { cdo ("-z zip_9 div",this_file,
+#                 " krapp_land_only.nc ", new_filename)}
+# }
+# 
+# # rename variables
+# month_prefix<-tolower(substr(month.name,1,3))
+# long_var_names<-c(biome4output="monthly net primary productivity",prec="monthly precipitation",
+#                   temp="monthly mean temperature",tcc="monthly total cloud cover")
+# var_names<-c(biome4output="mo_npp",prec="prec",
+#              temp="temp",tcc="tcc")
+# for (i in dir("./monthly/")){
+#   this_month_id<-unlist(strsplit(unlist(strsplit(i,"_"))[3],".",fixed=T))[1]
+#   this_var <- unlist(strsplit(i,"_"))[1]
+#   # update the variable name
+#   ncatted (paste0("-O -a long_name,",var_names[this_var],",o,c,'",month.name[as.numeric(this_month_id)],
+#                   " ",long_var_names[this_var],"' ",file.path("./monthly",i)))
+#   ncrename("-h -O -v ",csl(var_names[this_var],paste(var_names[this_var],this_month_id,sep="_")),file.path("./monthly",i))
+# }
+# 
+# # now combine monthly estimates for each variable
+# cdo ("-z zip_9", "merge","./monthly/biome*.nc","Krapp2021_npp_monthly_v1.1.0.nc")
+# cdo ("-z zip_9", "merge","./monthly/prec*.nc","Krapp2021_prec_monthly_v1.1.0.nc")
+# cdo ("-z zip_9", "merge","./monthly/temp*.nc","Krapp2021_temp_monthly_v1.1.0.nc")
+# cdo ("-z zip_9", "merge","./monthly/tcc*.nc","Krapp2021_tcc_monthly_v1.1.0.nc")
 
-# use the landmask to cut out the ice
-month_prefix<-tolower(substr(month.name,1,3))
-for (this_file in list.files(monthly_dirs, full.names = TRUE)){
-  i <- basename(this_file)
-  # get the month
-  this_month<-unlist(strsplit(unlist(strsplit(i,"_"))[3],".",fixed=T))[1]
-  this_month_id <- match(this_month,month_prefix)
-  if (this_month_id<10){
-    this_month_id <- paste0("0",this_month_id)
-  }
-  new_filename<-paste(unlist(strsplit(i,"_"))[1:2],collapse="_")
-  new_filename<-paste0(new_filename,"_",this_month_id,".nc")
-  new_filename<-file.path("./temp/monthly",new_filename)
-  # mask the file
-  if (unlist(strsplit(i,"_"))[1]=="biome4output"){
-    cdo ("-z zip_9 div -selname,mo_npp",this_file,
-         " krapp_land_only.nc ", new_filename)} else
-         { cdo ("-z zip_9 div",this_file,
-                " krapp_land_only.nc ", new_filename)}
-}
-
-# rename variables
-month_prefix<-tolower(substr(month.name,1,3))
-long_var_names<-c(biome4output="monthly net primary productivity",prec="monthly precipitation",
-                  temp="monthly mean temperature",tcc="monthly total cloud cover")
-var_names<-c(biome4output="mo_npp",prec="prec",
-             temp="temp",tcc="tcc")
-for (i in dir("./monthly/")){
-  this_month_id<-unlist(strsplit(unlist(strsplit(i,"_"))[3],".",fixed=T))[1]
-  this_var <- unlist(strsplit(i,"_"))[1]
-  # update the variable name
-  ncatted (paste0("-O -a long_name,",var_names[this_var],",o,c,'",month.name[as.numeric(this_month_id)],
-                  " ",long_var_names[this_var],"' ",file.path("./monthly",i)))
-  ncrename("-h -O -v ",csl(var_names[this_var],paste(var_names[this_var],this_month_id,sep="_")),file.path("./monthly",i))
-}
-
-# now combine monthly estimates for each variable
-cdo ("-z zip_9", "merge","./monthly/biome*.nc","Krapp2021_npp_monthly_v1.1.0.nc")
-cdo ("-z zip_9", "merge","./monthly/prec*.nc","Krapp2021_prec_monthly_v1.1.0.nc")
-cdo ("-z zip_9", "merge","./monthly/temp*.nc","Krapp2021_temp_monthly_v1.1.0.nc")
-cdo ("-z zip_9", "merge","./monthly/tcc*.nc","Krapp2021_tcc_monthly_v1.1.0.nc")
-
-
+## example commands for monthly variables to change units
+#cdo -z zip_9 -b F32 subc,273.15 Krapp2021_temp_monthly_v1.2.2.nc Krapp2021_temp_monthly_v1.4.0.nc
+#cdo -z zip_9 -b F32 divc,12. Krapp2021_prec_monthly_v1.2.2.nc Krapp2021_prec_monthly_v1.4.0.nc
+#ncatted -a pastclim_version,global,m,c,'1.4.0' Krapp2021_temp_monthly_v1.4.0.nc
+#ncatted -a pastclim_version,global,m,c,'1.4.0' Krapp2021_prec_monthly_v1.4.0.nc
 
 
 
@@ -163,3 +172,24 @@ for (file_id in list.files("./repackaged/cut", full.names=TRUE)){
   ncdf4::ncatt_put(nc_in,varid="time", attname = "axis",attval = "T")
   ncdf4::nc_close(nc_in)
 }
+
+# clean up the metadata for the variables based on the information from the pastclim table
+full_meta <- pastclim:::dataset_list_included
+sub_meta <- full_meta[full_meta$dataset == "Krapp2021",]
+for (i_ncfile in list.files("./repackaged/cut", full.names=TRUE)){
+  nc_in <- ncdf4::nc_open(i_ncfile, write = TRUE)
+  # update meta units
+  this_var_names <- names(nc_in$var)
+  for (x in this_var_names){
+    ncdf4::ncatt_put(nc_in, varid=x, attname = "long_name",
+                     attval = sub_meta$long_name[sub_meta$ncvar==x])
+    ncdf4::ncatt_put(nc_in, varid=x, attname = "units",
+                     attval = sub_meta$units[sub_meta$ncvar==x])
+  }
+  ncdf4::nc_close(nc_in)
+  # now remove the attribute unit if present (it should be units)
+  for (x in this_var_names){
+    ncatted(paste0('-a unit,',x,',d,, -h ',i_ncfile))
+  }
+}
+