+++ NEWS.md
@@ -1,7 +1,10 @@
 # development
-* Add Barreto et al 2023 (based on PALEO-PGEM, covering the last 5 M years),
-and all the WorldClim data (present, and future projections with multiple models
+* Add Barreto et al 2023 (based on PALEO-PGEM, covering the last 5 M years)
+* Add all the WorldClim data (present, and future projections with multiple models
 and emission scenarios).
+* Change the units of Krapp et al 2021 to match those of other datasets. Also, fix
+data duplication of some variables which has now also been fixed on the OSF repository
+for that dataset.
 * Improve `get_ice_mask()`, `get_land_mask()`, and `distance_from_sea()` to work
   on series rather than just on slices.
 * Speed up `region_*()` functions when subsetting the extent/cropping.